[fix] change package name domain to entity
