+++ edukit-core/src/main/java/com/edukit/member/entity/NicknameBannedWord.java
@@ -1,5 +1,6 @@
-package com.edukit.auth.domain;
+package com.edukit.member.entity;
 
+import com.edukit.common.domain.BaseTimeEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -12,16 +13,13 @@ import lombok.NoArgsConstructor;
 @Entity
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class ValidEmail {
+public class NicknameBannedWord extends BaseTimeEntity {
 
     @Id
-    @Column(name = "valid_email_id")
+    @Column(name = "nickname_banned_word_id")
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long id;
 
-    @Column(nullable = false, unique = true)
-    private String validEmail;
-
     @Column(nullable = false)
-    private String educationOffice;
+    private String word;
 }
