+++ edukit-core/src/main/java/com/edukit/member/entity/Subject.java
@@ -1,4 +1,4 @@
-package com.edukit.member.domain;
+package com.edukit.member.entity;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
