+++ edukit-core/src/main/java/com/edukit/post/entity/PostLike.java
@@ -1,6 +1,6 @@
-package com.edukit.post.domain;
+package com.edukit.post.entity;
 
-import com.edukit.member.domain.Member;
+import com.edukit.member.entity.Member;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
